[Add,Feat] 검색창 placeholder 애니메이션 hook 분리 (#1)
